requirements.txt
customtkinter
openai
aiohttp
Pillow
python-dotenv
cryptography
numpy
